% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KASPar_process.R
\name{KASPar_process}
\alias{KASPar_process}
\title{KASPar_process}
\usage{
KASPar_process(KASPar, output = TRUE)
}
\arguments{
\item{KASPar:}{string. The name of the input LightCycler480 .txt file}

\item{output:}{logical.  Do you want an output .pdf .log and .clean.csv file or to just print qc info to the console?}
}
\value{
If output is true, a log file, a PDF and a CSV with the cleaned data
}
\description{
Processes raw roche LightCycler 480 endpoint read genotype data and converts this into a csv file
and removes unknown, duplicate and .  It also produces a log file and a PDF containing scatter
plots for error checking.
}
\examples{
KASPar_process(KASPar="examples/SNP1.txt", output = FALSE)
}
\author{
Michael Way
}

