% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KASPar_annotate.R
\name{KASPar_annotate}
\alias{KASPar_annotate}
\title{Process a SNPs cleaned genotype data from csv format}
\usage{
KASPar_annotate(csv, RefMajor = "A2", RefMinor = "A1", rsid = "rs123456",
  direction = NA)
}
\arguments{
\item{csv}{The name of the input csv file}

\item{RefMajor}{- The reference major allele (e.g. A, G, T, C)}

\item{RefMinor}{- The reference minor allele (e.g. A, G, T, C)}

\item{rsid}{- The SNP identification number (is used to name the output)}

\item{direction}{- The direction of the allele codes (e.g. FWD, REV)}
}
\value{
A text file with genotype and allele annotations
KASPar_annotate()
}
\description{
Process a SNPs cleaned genotype data from csv format
}

