% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KASPar_plink.R
\name{KASPar_plink}
\alias{KASPar_plink}
\title{Converts several KASPar annotated  CSV files into a single PED file}
\usage{
KASPar_plink(name, samples)
}
\arguments{
\item{name}{- the prefix of the ped and map file}

\item{sample}{- a text file containing the first 6 lines of the ped file}
}
\value{
A PED file
KASPar_plink()
}
\description{
Converts several KASPar annotated  CSV files into a single PED file
}
\note{
This software requires a sample file with the first 6 columns of a ped file

and a map file with the names of variants
}

