% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KASPar_process.R
\name{KASPar_process}
\alias{KASPar_process}
\title{Processes raw roche LightCycler 480 endpoint read genotype data and converts this into a csv file
and removes unknown, duplicate and .  It also produces a log file and a PDF containing scatter
plots for error checking.}
\usage{
KASPar_process(KASPar)
}
\arguments{
\item{csv}{- The name of the input csv file}
}
\value{
A log file, a PDF and a CSV with the cleaned data
KASPar_process()
}
\description{
Processes raw roche LightCycler 480 endpoint read genotype data and converts this into a csv file
and removes unknown, duplicate and .  It also produces a log file and a PDF containing scatter
plots for error checking.
}

